OriginTrail 6.0.0-beta.2.2.5 Testnet Release

+++ src/service/file-service.js
@@ -1,6 +1,6 @@
-const path = require('path');
-const { mkdir, writeFile, readFile, unlink, stat } = require('fs/promises');
-const appRootPath = require('app-root-path');
+import path from 'path';
+import { mkdir, writeFile, readFile, unlink, stat, readdir } from 'fs/promises';
+import appRootPath from 'app-root-path';
 
 const MIGRATION_FOLDER_NAME = 'migrations';
 
@@ -31,6 +31,14 @@ class FileService {
 
     readFileOnPath(filePath) {
         return this._readFile(filePath, false);
+    }
+
+    async readDirectory(dirPath) {
+        return readdir(dirPath);
+    }
+
+    async stat(filePath) {
+        return stat(filePath);
     }
 
     /**
@@ -64,9 +72,13 @@ class FileService {
     }
 
     async removeFile(filePath) {
-        this.logger.debug(`Removing file on path: ${filePath}`);
-        await unlink(filePath);
-        return true;
+        if (await this.fileExists(filePath)) {
+            this.logger.trace(`Removing file on path: ${filePath}`);
+            await unlink(filePath);
+            return true;
+        }
+        this.logger.debug(`File doesn't exist on file path: ${filePath}`);
+        return false;
     }
 
     getDataFolderPath() {
@@ -91,6 +103,7 @@ class FileService {
     getOperationIdDocumentPath(operationId) {
         return path.join(this.getOperationIdCachePath(), operationId);
     }
+
 }
 
-module.exports = FileService;
+export default FileService;
